+ cancel buttons on selected balls

+++ app/src/main/java/kz/regto/database/d_entry_set.java
@@ -91,7 +91,7 @@ public class d_entry_set {
         return entry_value;
     }
 
-    public long getLog_id() {
+    public int getLog_id() {
         return log_id;
     }
